backend/requirements.txt
# Core Frameworks
Django==5.0.2
djangorestframework==3.14.0
python-dotenv==1.0.1
django-cors-headers==4.3.1

# AI & LLM (Google Gemini + LangChain)
google-generativeai>=0.8.3
langchain>=0.3.7
langchain-google-genai>=2.0.4
langchain-community>=0.3.7
langchain-core>=0.3.15
langchain-text-splitters>=0.3.2

# Database & Vector Store
psycopg2-binary==2.9.9
pgvector==0.3.6

# Utilities
requests==2.32.5
pypdf==5.1.0
tiktoken==0.8.0
numpy==1.26.4